Update GitHub Actions to Use crankkio/web Repository (#9)
## Description

This PR updates the GitHub Actions workflows to pull the web UI from the
crankkio/web repository instead of the meshtastic/web repository. The
changes affect the following workflow files:

- build_esp32.yml

- build_esp32_c3.yml

- build_esp32_s3.yml

- package_amd64.yml

These updates ensure that the correct repository is used for fetching
the web UI assets during the build and packaging processes.